pyproject.toml
[project]
name = "returns-portal-challenge"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "django>=6.0,<7.0",
  "django-pydantic-field>=0.5,<1.0",
  "drf-spectacular>=0.28,<1.0",
  "djangorestframework>=3.0,<4.0",
  "pydantic>=2.0,<3.0",
  "pyyaml>=6.0,<7.0",
]

[dependency-groups]
dev = [
  "django-stubs>=5.0,<6.0",
  "pre-commit>=4.0,<5.0",
  "pytest>=9.0,<10.0",
  "pytest-django>=4.0,<5.0",
  "ruff>=0.15,<1.0",
  "mypy>=1.0,<2.0",
]

[tool.uv]
default-groups = ["dev"]

[tool.pytest.ini_options]
testpaths = ["**/tests"]
pythonpath = ["."]
addopts = "-q"
DJANGO_SETTINGS_MODULE = "returns_portal.settings"


[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
strict = true
ignore_missing_imports = true
plugins = ["mypy_django_plugin.main", "pydantic.mypy"]

[tool.django-stubs]
django_settings_module = "returns_portal.settings"
